fix: 4 code-review issues in orchestrator/results/historical/goals
- historical.py: fix IndexError crash when sequence_of_returns_risk's
  worst_n equals n_years (glide-path per-year composition always hits
  this), by typing the empty rest_idx array as int.
- results.py: guard per-period-return division against 0/0 -> NaN when
  many paths are ruined by an aggressive cashflow, so percentile_table,
  withdrawal_rates_by_percentile, and sharpe_sortino_by_percentile never
  leak NaN into the JSON response.
- orchestrator.py: correct the displayed glide-path allocation schedule
  to match frontend/src/api/mockData.ts's shipped semantics (hold start
  allocation until years_to_retirement - glide_path_years, complete
  transition exactly at years_to_retirement) instead of starting the
  transition at year 0.
- goals.py: apply_cashflow now scales amount by cashflow_frequency,
  matching apply_named_goals's existing frequency scaling, so a monthly
  cashflow actually applies 12x an annual one instead of being silently
  ignored.

+++ backend/app/engine/historical.py
@@ -69,7 +69,7 @@ def _apply_sequence_of_returns_risk(sampled: np.ndarray, worst_n: int) -> np.nda
     for p in range(sampled.shape[0]):
         row = sampled[p]
         worst_idx = np.argsort(row)[:worst_n]
-        rest_idx = np.array([i for i in range(n_years) if i not in set(worst_idx)])
+        rest_idx = np.array([i for i in range(n_years) if i not in set(worst_idx)], dtype=int)
         worst_sorted = row[worst_idx][np.argsort(row[worst_idx])]
         reordered[p] = np.concatenate([worst_sorted, row[rest_idx]])
     return reordered
